+++ module/sheets/torgeternityActorSheet.js
@@ -441,6 +441,7 @@ export default class torgeternityActorSheet extends ActorSheet {
       customSkill: event.currentTarget.dataset.customskill,
       actor: this.actor.uuid,
       actorPic: this.actor.img,
+      actorName: this.actor.name,
       actorType: this.actor.type,
       isAttack: false,
       isFav: this.actor.system.skills[skillName]?.isFav || this.actor.system.attributes?.[skillName + "IsFav"] || isFav,
@@ -469,6 +470,7 @@ export default class torgeternityActorSheet extends ActorSheet {
       customSkill: "false",
       actor: this.actor.uuid,
       actorPic: this.actor.img,
+      actorName: this.actor.name,
       actorType: "vehicle",
       isAttack: false,
       skillName: "Vehicle Chase",
@@ -493,6 +495,7 @@ export default class torgeternityActorSheet extends ActorSheet {
       customSkill: "false",
       actor: this.actor.uuid,
       actorPic: this.actor.img,
+      actorName: this.actor.name,
       actorType: "vehicle",
       isAttack: false,
       skillName: "Vehicle Operation",
@@ -525,6 +528,7 @@ export default class torgeternityActorSheet extends ActorSheet {
       customSkill: "false",
       actor: this.actor.uuid,
       actorPic: this.actor.img,
+      actorName: this.actor.name,
       actorType: "vehicle",
       isAttack: false,
       skillName: "Vehicle Stunt",
@@ -571,6 +575,7 @@ export default class torgeternityActorSheet extends ActorSheet {
       testType: "interactionAttack",
       actor: this.actor.uuid,
       actorPic: this.actor.img,
+      actorName: this.actor.name,
       actorType: this.actor.type,
       isAttack: false,
       interactionAttackType: event.currentTarget.getAttribute("data-attack-type"),
@@ -617,6 +622,7 @@ export default class torgeternityActorSheet extends ActorSheet {
       testType: "attack",
       actor: this.actor.uuid,
       actorPic: this.actor.img,
+      actorName: this.actor.name,
       actorType: this.actor.type,
       amountBD: 0,
       attackType: "unarmedCombat",
@@ -668,6 +674,7 @@ export default class torgeternityActorSheet extends ActorSheet {
       activelyDefending: false,
       actor: this.actor.uuid,
       actorPic: this.actor.img,
+      actorName: this.actor.name,
       actorType: this.actor.type,
       isAttack: false,
       skillName: "activeDefense",
@@ -701,6 +708,7 @@ export default class torgeternityActorSheet extends ActorSheet {
       activelyDefending: true,
       actor: this.actor.uuid,
       actorPic: this.actor.img,
+      actorName: this.actor.name,
       actorType: this.actor.type,
       isAttack: false,
       skillName: "activeDefense",
@@ -815,6 +823,7 @@ export default class torgeternityActorSheet extends ActorSheet {
       testType: "attack",
       actor: this.actor.uuid,
       actorPic: this.actor.img,
+      actorName: this.actor.name,
       actorType: this.actor.type,
       amountBD: 0,
       attackType: attackType,
@@ -900,6 +909,7 @@ export default class torgeternityActorSheet extends ActorSheet {
       testType: "power",
       actor: this.actor.uuid,
       actorPic: this.actor.img,
+      actorName: this.actor.name,
       actorType: this.actor.type,
       attackType: "power",
       powerName: item.name,
@@ -983,7 +993,8 @@ export function checkUnskilled(skillValue, skillName, actor) {
 
     let templateData = {
       message: skillName + " " + game.i18n.localize("torgeternity.chatText.check.cantUseUntrained"),
-      actorPic: actor.system.img,
+      actorPic: actor.img,
+      actorName: actor.name,
     };
 
     const templatePromise = renderTemplate(
