+++ templates/partials/activeDefense-card.hbs
@@ -1,12 +1,9 @@
-<div
-  class="skill-card sheet-card"
-  data-type="{{data.type}}"
-  style="background-image: url('systems/torgeternity/images/bgparch.webp')"
->
+<div class="skill-card sheet-card" data-type="{{data.type}}"
+  style="background-image: url('systems/torgeternity/images/bgparch.webp')">
   <table style="border:none;background-color:transparent;margin:2px 0px 0px 0px">
     <tr style="border:none">
       <td>
-        <img class="actor-icon" src="{{actorPic}}" title="{{name}}"/>
+        <img class="actor-icon" src="{{actorPic}}" title="{{actorName}}" />
       </td>
       <td class="chat-title" style="font-size:20px;padding:0px 0px 0px 0px">
         {{localize "torgeternity.sheetLabels.activeDefense"}}
