+++ templates/partials/possibility-card.hbs
@@ -1,12 +1,9 @@
-<div
-  class="possibility-card sheet-card"
-  data-type="{{data.type}}"
-  style="background-image: url('systems/torgeternity/images/bgparch.webp')"
->
+<div class="possibility-card sheet-card" data-type="{{data.type}}"
+  style="background-image: url('systems/torgeternity/images/bgparch.webp')">
   <table style="border:none;background-color:transparent">
     <tr style="border:none">
       <td>
-        <img class="actor-icon" src="{{actorPic}}" title="{{name}}"/>
+        <img class="actor-icon" src="{{actorPic}}" title="{{actorName}}" />
       </td>
       <td class="chat-title" style="font-size:20px">
         {{localize "torgeternity.chatText.possibility"}}
