+++ templates/partials/skill-card.hbs
@@ -6,7 +6,7 @@
   <table style="border:none;background-color:transparent;margin:2px 0px 0px 0px">
     <tr style="border:none">
       <td>
-        <img class="actor-icon" src="{{actorPic}}" title="{{name}}" />
+        <img class="actor-icon" src="{{actorPic}}" title="{{actorName}}" />
         {{log this}}
       </td>
       <td>
