+++ templates/partials/skill-error-card.hbs
@@ -6,7 +6,7 @@
   <table style="border:none;background-color:transparent;margin:2px 0px 0px 0px">
     <tr style="border:none">
       <td>
-        <img class="actor-icon" src="{{actorPic}}" title="{{name}}"/>
+        <img class="actor-icon" src="{{actorPic}}" title="{{actorName}}" />
       </td>
       <td class="chat-title" style="font-size:18px;padding:0px 0px 0px 0px">
         {{localize "torgeternity.chatText.check.testUnsuccessfull"}}
