+++ templates/partials/up-card.hbs
@@ -6,7 +6,7 @@
   <table style="border:none;background-color:transparent">
     <tr style="border:none">
       <td style="width:52px">
-        <img class="actor-icon" src="{{actorPic}}" title="{{name}}"/>
+        <img class="actor-icon" src="{{actorPic}}" title="{{actorName}}" />
       </td>
       <td class="chat-title" style="font-size:20px">
         {{localize "torgeternity.chatText.up"}}
