+++ torgeternity.js
@@ -651,6 +651,7 @@ function rollItemMacro(itemName) {
         amountBD: 0,
         isFav: skillData.isFav,
         actorPic: actor.img,
+        actorName: actor.name,
         skillName: weaponData.attackWith,
         skillBaseAttribute: skillData.baseAttribute,
         skillValue: skillData.value,
@@ -724,6 +725,7 @@ function rollItemMacro(itemName) {
         testType: "power",
         actor: actor.uuid,
         actorPic: actor.img,
+        actorName: actor.name,
         actorType: actor.type,
         attackType: "power",
         powerName: item.name,
@@ -838,6 +840,7 @@ function rollSkillMacro(skillName, attributeName, isInteractionAttack) {
     testType: isAttributeTest ? "attribute" : "skill",
     actor: actor.uuid,
     actorPic: actor.img,
+    actorName: actor.name,
     actorType: actor.type,
     skillName: isAttributeTest ? attributeName : skillName,
     skillBaseAttribute: game.i18n.localize("torgeternity.attributes." + attributeName),
